requirements.txt
pyyaml
torch
torchaudio
numpy
Pillow
ultralytics
matplotlib
soundfile
pydub
chardet
scikit-image